+++ R/datasets.r
@@ -1,7 +1,7 @@
 #' List the datasets in a project
 #'
-#' @param project The project name, a string
 #' @return a character vector of dataset names
+#' @inheritParams list_tables
 #' @seealso Google API documentation:
 #'   \url{https://developers.google.com/bigquery/docs/reference/v2/datasets/list}
 #' @family datasets
@@ -11,13 +11,17 @@
 #' list_datasets("publicdata")
 #' list_datasets("githubarchive")
 #' }
-list_datasets <- function(project) {
+list_datasets <- function(project, page_size = 50, max_pages = Inf) {
   assert_that(is.string(project))
 
-  url <- sprintf("projects/%s/datasets", project)
-  data <- bq_get(url)$datasets
+  pages <- bq_get_paginated(
+    bq_path(project, ""),
+    page_size = page_size,
+    max_pages = max_pages
+  )
 
-  unlist(lapply(data, function(x) x$datasetReference$datasetId))
+  datasets <- unlist(lapply(pages, function(x) x$datasets), recursive = FALSE)
+  vapply(datasets, function(x) x$datasetReference$datasetId, character(1))
 }
 
 #' Gets an existing dataset in a project
@@ -34,10 +38,18 @@ list_datasets <- function(project) {
 #' get_dataset("publicdata", "shakespeare")
 #' }
 get_dataset <- function(project, dataset) {
-  assert_that(is.string(project), is.string(dataset))
+  bq_get(bq_path(project, dataset))
+}
 
-  url <- sprintf("projects/%s/datasets/%s", project, dataset)
-  bq_get(url)
+#' @rdname get_dataset
+#' @export
+#' @description `exists_dataset` merely checks if a table exists, and returns
+#'   either `TRUE` or `FALSE`.
+exists_dataset <- function(project, dataset) {
+  tryCatch(
+    !is.null(get_dataset(project = project, dataset = dataset)),
+    bigrquery_notFound = function(e) FALSE
+  )
 }
 
 #' Deletes an existing dataset in a project
@@ -65,8 +77,9 @@ delete_dataset <- function(project, dataset, deleteContents = FALSE) {
 #'
 #' @param project The project name, a string
 #' @param dataset The name of the dataset to create, a string
-#' @param description The dataset description, a string
-#' @param friendlyName The dataset's friendly name, a string
+#' @param ... Additional arguments merged into the body of the
+#'   request. `snake_case` will automatically be converted into
+#'   `camelCase` so you can use consistent argument names.
 #' @seealso Google API documentation:
 #'   \url{https://cloud.google.com/bigquery/docs/reference/v2/datasets/insert}
 #' @family datasets
@@ -75,29 +88,23 @@ delete_dataset <- function(project, dataset, deleteContents = FALSE) {
 #' \dontrun{
 #' insert_dataset("myproject", "new_dataset")
 #' }
-insert_dataset <- function(project, dataset, description = NULL, friendlyName = NULL) {
+insert_dataset <- function(project, dataset, ...) {
   assert_that(is.string(project), is.string(dataset))
 
-  url <- sprintf("projects/%s/datasets", project)
-
-  body = list(
+  url <- bq_path(project, "")
+  body <- list(
     datasetReference = list(
       projectId = project,
       datasetId = dataset
-    ),
-    description = description,
-    friendlyName = friendlyName
+    )
   )
 
-  bq_post(url, body)
+  bq_post(url, body = bq_body(body, ...))
 }
 
 #' Updates an existing dataset in a project
 #'
-#' @param project The project name, a string
-#' @param dataset The dataset to update, a string
-#' @param description The dataset description, a string
-#' @param friendlyName The dataset's friendly name, a string
+#' @inheritParams insert_dataset
 #' @seealso Google API documentation:
 #'   \url{https://cloud.google.com/bigquery/docs/reference/v2/datasets/update}
 #' @family datasets
@@ -106,19 +113,14 @@ insert_dataset <- function(project, dataset, description = NULL, friendlyName = 
 #' \dontrun{
 #' update_dataset("myproject", "existing_dataset", "my description", "friendly name")
 #' }
-update_dataset <- function(project, dataset, description = NULL, friendlyName = NULL) {
-  assert_that(is.string(project), is.string(dataset))
-
-  url <- sprintf("projects/%s/datasets/%s", project, dataset)
-
-  body = list(
+update_dataset <- function(project, dataset, ...) {
+  url <- bq_path(project, dataset)
+  body <- list(
     datasetReference = list(
       projectId = project,
       datasetId = dataset
-    ),
-    description = description,
-    friendlyName = friendlyName
+    )
   )
 
-  bq_put(url, body)
+  bq_put(url, body = bq_body(body, ...))
 }