+++ R/job-upload.r
@@ -8,15 +8,16 @@
 #' @param values data frame of data to upload
 #' @param billing project ID to use for billing
 #' @param create_disposition behavior for table creation if the destination
-#'   already exists. defaults to \code{"CREATE_IF_NEEDED"},
-#'   the only other supported value is \code{"CREATE_NEVER"}; see
+#'   already exists. defaults to `"CREATE_IF_NEEDED"`,
+#'   the only other supported value is `"CREATE_NEVER"`; see
 #'   \href{https://cloud.google.com/bigquery/docs/reference/v2/jobs#configuration.load.createDisposition}{the API documentation}
 #'   for more information
 #' @param write_disposition behavior for writing data if the destination already
-#'   exists. defaults to \code{"WRITE_APPEND"}, other possible values are
-#'   \code{"WRITE_TRUNCATE"} and \code{"WRITE_EMPTY"}; see
+#'   exists. defaults to `"WRITE_APPEND"`, other possible values are
+#'   `"WRITE_TRUNCATE"` and `"WRITE_EMPTY"`; see
 #'   \href{https://cloud.google.com/bigquery/docs/reference/v2/jobs#configuration.load.writeDisposition}{the API documentation}
 #'   for more information
+#' @inheritParams insert_dataset
 #' @seealso Google API documentation:
 #' \url{https://developers.google.com/bigquery/loading-data-into-bigquery#loaddatapostrequest}
 #' @family jobs
@@ -33,15 +34,21 @@
 insert_upload_job <- function(project, dataset, table, values,
                               billing = project,
                               create_disposition = "CREATE_IF_NEEDED",
-                              write_disposition = "WRITE_APPEND") {
-  assert_that(is.string(project), is.string(dataset), is.string(table),
-    is.data.frame(values), is.string(billing))
+                              write_disposition = "WRITE_APPEND",
+                              ...) {
+  assert_that(
+    is.string(project),
+    is.string(dataset),
+    is.string(table),
+    is.data.frame(values),
+    is.string(billing)
+  )
 
   # https://developers.google.com/bigquery/docs/reference/v2/jobs#resource
   config <- list(
     configuration = list(
       load = list(
-        sourceFormat = "CSV",
+        sourceFormat = "NEWLINE_DELIMITED_JSON",
         schema = list(
           fields = schema_fields(values)
         ),
@@ -55,14 +62,18 @@ insert_upload_job <- function(project, dataset, table, values,
       )
     )
   )
-  config_part <- part(c("Content-type" = "application/json; charset=UTF-8"),
-    jsonlite::toJSON(config, pretty = TRUE))
+  config <- bq_body(config, ...)
+  config_part <- part(
+    c("Content-type" = "application/json; charset=UTF-8"),
+    jsonlite::toJSON(config, auto_unbox = TRUE)
+  )
+  data_part <- part(
+    c("Content-type" = "application/json; charset=UTF-8"),
+    export_json(values)
+  )
 
-  csv <- standard_csv(values)
-  csv_part <- part(c("Content-type" = "application/octet-stream"), csv)
-
-  url <- sprintf("projects/%s/jobs", billing)
-  bq_upload(url, c(config_part, csv_part))
+  url <- bq_path(billing, jobs = "")
+  bq_upload(url, c(config_part, data_part))
 }
 
 schema_fields <- function(data) {
@@ -73,41 +84,30 @@ schema_fields <- function(data) {
 data_type <- function(x) {
   if (is.factor(x)) return("STRING")
   if (inherits(x, "POSIXt")) return("TIMESTAMP")
-  if (inherits(x, "Date")) return("TIMESTAMP")
+  if (inherits(x, "hms")) return("TIME")
+  if (inherits(x, "Date")) return("DATE")
 
-  switch(typeof(x),
-         character = "STRING",
-         logical = "BOOLEAN",
-         double = "FLOAT",
-         integer = "INTEGER",
-         stop("Unsupported type: ", typeof(x), call. = FALSE)
+  switch(
+    typeof(x),
+    character = "STRING",
+    logical = "BOOLEAN",
+    double = "FLOAT",
+    integer = "INTEGER",
+    stop("Unsupported type: ", typeof(x), call. = FALSE)
   )
 }
 
-standard_csv <- function(values) {
-  # Convert factors to strings
-  is_factor <- vapply(values, is.factor, logical(1))
-  values[is_factor] <- lapply(values[is_factor], as.character)
+export_json <- function(values) {
+  # Eliminate row names
+  rownames(values) <- NULL
 
-  # Encode special characters in strings
-  is_char <- vapply(values, is.character, logical(1))
-  values[is_char] <- lapply(values[is_char], encodeString, na.encode = FALSE)
-
-  # Encode dates and times
-  is_time <- vapply(values, function(x) inherits(x, "POSIXct"), logical(1))
+  # Convert times to unix timestamps
+  is_time <- vapply(values, function(x) inherits(x, "POSIXt"), logical(1))
   values[is_time] <- lapply(values[is_time], as.numeric)
 
-  is_date <- vapply(values, function(x) inherits(x, "Date"), logical(1))
-  values[is_date] <- lapply(values[is_date], function(x) as.numeric(as.POSIXct(x)))
+  con <- rawConnection(raw(0), "r+")
+  on.exit(close(con))
+  jsonlite::stream_out(values, con, verbose = FALSE, na = "null")
 
-  tmp <- tempfile(fileext = ".csv")
-  on.exit(unlink(tmp))
-
-  conn <- file(tmp, open = "wb")
-  utils::write.table(values, conn, sep = ",", na = "", qmethod = "double",
-              row.names = FALSE, col.names = FALSE, eol = "\12")
-  close(conn)
-
-  # Don't read trailing nl
-  readChar(tmp, file.info(tmp)$size - 1, useBytes = TRUE)
+  rawToChar(rawConnectionValue(con))
 }