+++ R/request.r
@@ -11,43 +11,127 @@ prepare_bq_query <- function(query) {
   query
 }
 
+bq_path <- function(project, dataset = NULL, table = NULL, ...) {
+  assert_that(is.null(project) || is.string(project))
+  assert_that(is.null(table) || is.string(table))
+  assert_that(is.null(dataset) || is.string(dataset))
+
+  components <- c(
+    projects = project,
+    datasets = dataset,
+    tables = table,
+    ...
+  )
+
+  paste0(names(components), "/", components, collapse = "/")
+}
+
+bq_ua <- function() {
+  httr::user_agent(paste0(
+    "bigrquery/", utils::packageVersion("bigrquery"), " ",
+    "httr/", utils::packageVersion("httr")
+  ))
+}
+
+bq_body <- function(body, ...) {
+  user <- toCamel(list(...))
+  utils::modifyList(body, user)
+}
+
+
 #' @importFrom httr GET config
 bq_get <- function(url, ..., query = NULL, token = get_access_cred()) {
-  req <- GET(paste0(base_url, url), config(token = token), ...,
-             query = prepare_bq_query(query))
+  req <- GET(
+    paste0(base_url, url),
+    config(token = token),
+    bq_ua(),
+    ...,
+    query = prepare_bq_query(query)
+  )
   process_request(req)
 }
 
+#' @importFrom httr GET config
+bq_get_paginated <- function(url, ..., query = NULL, token = get_access_cred(),
+                             page_size = 50, max_pages = Inf) {
+
+  assert_that(is.numeric(max_pages), length(max_pages) == 1)
+  assert_that(is.numeric(page_size), length(page_size) == 1)
+
+  query <- utils::modifyList(list(maxResults = page_size), query %||% list())
+  pages <- list()
+
+  page <- bq_get(url, ..., query = query, token = token)
+  i <- 1
+  pages[[i]] <- page
+  page_token <- page$nextPageToken
+
+  while (!is.null(page_token) && i < max_pages) {
+    query$pageToken <- page_token
+    page <- bq_get(url, ..., query = query, token = token)
+
+    i <- i + 1
+    pages[[i]] <- page
+    page_token <- page$nextPageToken
+  }
+
+  pages
+}
+
+
 #' @importFrom httr DELETE config
 bq_delete <- function(url, ..., query = NULL, token = get_access_cred()) {
-  req <- DELETE(paste0(base_url, url), config(token = token), ...,
-                query = prepare_bq_query(query))
+  req <- DELETE(
+    paste0(base_url, url),
+    config(token = token),
+    bq_ua(),
+    ...,
+    query = prepare_bq_query(query)
+  )
   process_request(req)
 }
 
 #' @importFrom httr POST add_headers config
 bq_post <- function(url, body, ..., query = NULL, token = get_access_cred()) {
   json <- jsonlite::toJSON(body)
-  req <- POST(paste0(base_url, url), body = json, config(token = token),
-              add_headers("Content-Type" = "application/json"), ...,
-              query = prepare_bq_query(query))
-  process_request(req)
+  req <- POST(
+    paste0(base_url, url),
+    body = json,
+    bq_ua(),
+    config(token = token),
+    add_headers("Content-Type" = "application/json"),
+    ...,
+    query = prepare_bq_query(query)
+  )
+  invisible(process_request(req))
 }
 
 #' @importFrom httr PUT add_headers config
 bq_put <- function(url, body, ..., query = NULL, token = get_access_cred()) {
   json <- jsonlite::toJSON(body)
-  req <- PUT(paste0(base_url, url), body = json, config(token = token),
-              add_headers("Content-Type" = "application/json"), ...,
-              query = prepare_bq_query(query))
+  req <- PUT(
+    paste0(base_url, url),
+    body = json,
+    bq_ua(),
+    config(token = token),
+    add_headers("Content-Type" = "application/json"),
+    ...,
+    query = prepare_bq_query(query)
+  )
   process_request(req)
 }
 
 #' @importFrom httr POST add_headers config
 bq_upload <- function(url, parts, ..., query = NULL, token = get_access_cred()) {
   url <- paste0(upload_url, url)
-  req <- POST_multipart_related(url, parts = parts, config(token = token), ...,
-                                query = prepare_bq_query(query))
+  req <- POST_multipart_related(
+    url,
+    parts = parts,
+    config(token = token),
+    bq_ua(),
+    ...,
+    query = prepare_bq_query(query)
+  )
   process_request(req)
 }