+++ R/tabledata.r
@@ -1,20 +1,19 @@
 #' Retrieve data from a table.
 #'
-#' \code{list_tabledata} returns a single dataframe.
-#'
 #' @inheritParams get_table
 #' @param callback function called with single argument, the data from the
 #'   current page of data
-#' @param quiet if \code{FALSE}, prints informative status messages.
+#' @param quiet if `FALSE`, prints informative status messages.
 #' @param table_info if known, the table information retrieved with
-#'   \code{\link{get_table}}
+#'   [get_table()]
 #' @param page_size Number of items per page.
-#' @param warn If \code{TRUE}, warn when there are rows remaining to
+#' @param warn If `TRUE`, warn when there are rows remaining to
 #'   be pulled down from database.
-#' @param max_pages maximum number of pages to retrieve. Use \code{Inf}
+#' @param max_pages maximum number of pages to retrieve. Use `Inf`
 #'  to retrieve the complete dataset.
 #' @seealso API documentation at
 #'   \url{https://developers.google.com/bigquery/docs/reference/v2/tabledata/list}
+#' @return `list_tabledata` returns a single dataframe.
 #' @export
 #' @examples
 #' \dontrun{
@@ -23,9 +22,13 @@
 #'   page_size = 10)
 #' dim(natal)
 #' }
-list_tabledata <- function(project, dataset, table, page_size = 1e4,
-                           table_info = NULL, max_pages = 10, warn = TRUE,
-                           quiet = getOption("bigrquery.quiet")) {
+list_tabledata <- function(project, dataset, table,
+                           page_size = 1e4,
+                           table_info = NULL,
+                           max_pages = 10,
+                           warn = TRUE,
+                           quiet = getOption("bigrquery.quiet")
+                           ) {
   assert_that(is.string(project), is.string(dataset), is.string(table))
   assert_that(is.numeric(max_pages), length(max_pages) == 1, max_pages >= 1)
 
@@ -37,16 +40,23 @@ list_tabledata <- function(project, dataset, table, page_size = 1e4,
     rows <<- c(rows, list(new_rows))
   }
 
-  list_tabledata_callback(project, dataset, table, append_rows,
-    table_info = table_info, page_size = page_size, max_pages = max_pages,
-    warn = warn, quiet = quiet
+  list_tabledata_callback(
+    project,
+    dataset,
+    table,
+    append_rows,
+    table_info = table_info,
+    page_size = page_size,
+    max_pages = max_pages,
+    warn = warn,
+    quiet = quiet
   )
 
   do.call("rbind", rows)
 }
 
 #' @description
-#' \code{list_tabledata_callback} calls the supplied callback with each page
+#' `list_tabledata_callback` calls the supplied callback with each page
 #' of data.
 #' @rdname list_tabledata
 #' @export
@@ -60,31 +70,29 @@ list_tabledata_callback <- function(project, dataset, table, callback,
   assert_that(is.function(callback))
   assert_that(is.numeric(max_pages), length(max_pages) == 1, max_pages >= 1)
 
-  elapsed <- timer()
-  is_quiet <- function(x) isTRUE(quiet) || (is.na(quiet) && elapsed() < 2)
-
   iter <- list_tabledata_iter(
-    project = project, dataset = dataset, table = table,
-    table_info = table_info)
+    project = project,
+    dataset = dataset,
+    table = table,
+    table_info = table_info
+  )
+
+  pages <- min(ceiling(iter$get_rows() / page_size), max_pages)
+  progress <- bq_progress(
+    "Retrieving data [:bar] :percent eta: :eta",
+    total = pages,
+    quiet = quiet
+  )
 
   cur_page <- 0L
 
-  while(cur_page < max_pages && !iter$is_complete()) {
-    if (!is_quiet()) {
-      if (cur_page >= 1L) {
-        cat("\rRetrieving data: ", sprintf("%4.1f", elapsed()), "s", sep = "")
-      } else {
-        cat("Retrieving data")
-      }
-    }
-
+  while (cur_page < max_pages && !iter$is_complete()) {
     data <- iter$next_(page_size)
     callback(data)
 
+    progress$tick()
     cur_page <- cur_page + 1
   }
-
-  if (!is_quiet()) cat("\n")
 
   if (isTRUE(warn) && !iter$is_complete()) {
     warning("Only first ", max_pages, " pages of size ", page_size,
@@ -101,8 +109,7 @@ list_tabledata_iter <- function(project, dataset, table, table_info = NULL) {
   table_info <- table_info %||% get_table(project, dataset, table)
   schema <- table_info$schema
 
-  url <- sprintf("projects/%s/datasets/%s/tables/%s/data", project, dataset,
-    table)
+  url <- bq_path(project, dataset = dataset, table = table, data = "")
 
   last_response <- NULL
   rows_fetched <- 0
@@ -153,24 +160,50 @@ list_tabledata_iter <- function(project, dataset, table, table_info = NULL) {
     rows_fetched
   }
 
-  #' @description
-  #' \code{list_tabledata_iter} returns a named list with functions \code{next_}
-  #' (fetches one chunk of rows), \code{next_paged} (fetches arbitrarily many
-  #' rows using a specified page size), \code{is_complete} (checks if all rows
-  #' have been fetched), \code{get_schema} (returns the schema of the table),
-  #' and \code{get_rows_fetched} (returns the number of rows already fetched).
-  list(next_ = next_, next_paged = next_paged, is_complete = is_complete,
-       get_schema = get_schema, get_rows_fetched = get_rows_fetched)
+  get_rows <- function() {
+    as.numeric(table_info$numRows)
+  }
+
+  #' @return
+  #' `list_tabledata_iter` returns a named list with functions
+  #' * `next_` (fetches one chunk of rows)
+  #' * `next_paged` (fetches arbitrarily many rows using a specified page size)
+  #' * `is_complete` (checks if all rows have been fetched)
+  #' * `get_schema` (returns the schema of the table),
+  #' * `get_rows_fetched` (returns the number of rows already fetched).
+  #' * `get_rows` (returns total number of rows)
+  list(
+    next_ = next_,
+    next_paged = next_paged,
+    is_complete = is_complete,
+    get_schema = get_schema,
+    get_rows = get_rows,
+    get_rows_fetched = get_rows_fetched
+  )
 }
 
 #Types can be loaded into R, record is not supported yet.
-converter <- list(
-  integer = as.integer,
-  float = as.double,
-  boolean = as.logical,
-  string = identity,
-  timestamp = function(x) as.POSIXct(as.integer(x), origin = "1970-01-01", tz = "UTC")
-)
+converters <- function() {
+  list(
+    integer = as.integer,
+    float = as.double,
+    boolean = as.logical,
+    string = identity,
+    timestamp = function(x) {
+      as.POSIXct(as.numeric(x), origin = "1970-01-01", tz = "UTC")
+    },
+    time = function(x) {
+      readr::parse_time(x, "%H:%M:%OS")
+    },
+    date = function(x) {
+      readr::parse_date(x, format = "%Y-%m-%d")
+    },
+    datetime = function(x) {
+      readr::parse_datetime(x)
+    }
+  )
+}
+
 
 extract_data <- function(rows, schema) {
   if (is.null(rows) || length(rows) == 0L) {
@@ -184,17 +217,13 @@ extract_data <- function(rows, schema) {
   types <- tolower(vapply(schema$fields, function(x) x$type, character(1)))
 
   # Convert NULLs into NAs
-  out <- character(length(rows) * length(types))
-  for(i in seq_along(rows)) {
-    for(j in seq_along(types)) {
-      if (is.null(rows[[i]]$f[[j]]$v)) rows[[i]]$f[[j]]$v <- NA_character_
-    }
-  }
+  rows <- null_to_na(rows)
   data <- unlist(rows, use.names = FALSE)
   data_m <- matrix(data, nrow = length(types))
 
   out <- vector("list", length(types))
-  for(i in seq_along(types)) {
+  converter <- converters()
+  for (i in seq_along(types)) {
     type <- types[[i]]
     if (!(type %in% names(converter))) {
       stop("Don't know how to convert type ", type, call. = FALSE)
@@ -205,3 +234,10 @@ extract_data <- function(rows, schema) {
   names(out) <- vapply(schema$fields, function(x) x$name, character(1))
   as_df(out)
 }
+
+
+null_to_na <- function(x) {
+  .Call(null_to_na_, x)
+}
+
+