+++ R/tables.r
@@ -1,8 +1,8 @@
 #' List available tables in dataset.
 #'
 #' @inheritParams get_table
-#' @param max_results (optional) Maximum number of results to
-#'   retrieve.
+#' @param page_size Number of items per page
+#' @param max_pages Maximum number of pages to retrieve
 #' @return a character vector of table names
 #' @family tables
 #' @seealso API documentation:
@@ -12,29 +12,42 @@
 #' \dontrun{
 #' list_tables("publicdata", "samples")
 #' list_tables("githubarchive", "github")
-#' list_tables("publicdata", "samples", max_results = 2)
+#' list_tables("publicdata", "samples", max_pages = 2, page_size = 2)
 #' }
-list_tables <- function(project, dataset, max_results = NULL) {
-  assert_that(is.string(project), is.string(dataset))
-  if (!is.null(max_results)) {
-    assert_that(is.numeric(max_results), length(max_results) == 1)
-  }
+list_tables <- function(project, dataset, page_size = 50, max_pages = Inf) {
+  data <- bq_get_paginated(
+    bq_path(project, dataset, ""),
+    page_size = page_size,
+    max_pages = max_pages
+  )
 
-  url <- sprintf("projects/%s/datasets/%s/tables", project, dataset)
-  query <- list()
-  if (!is.null(max_results)) {
-    query$maxResults <- max_results
-  }
-  data <- bq_get(url, query = query)$tables
-  do.call("rbind", lapply(data, as.data.frame, row.names = 1L))
+  tables <- unlist(lapply(data, function(x) x$tables), recursive = FALSE)
+  vapply(tables, function(x) x$tableReference$tableId, character(1L))
+}
 
-  vapply(data, function(x) x$tableReference$tableId, character(1L))
+#' Insert empty table
+#'
+#' @inheritParams insert_dataset
+#' @inheritParams get_table
+#' @export
+#' @seealso API documentation:
+#'  \url{https://developers.google.com/bigquery/docs/reference/v2/tables/insert}
+insert_table <- function(project, dataset, table, ...) {
+  url <- bq_path(project, dataset, "")
+  body <- list(
+    tableReference = list(
+      projectId = project,
+      datasetId = dataset,
+      tableId = table
+    )
+  )
+
+  bq_post(url, body = bq_body(body, ...))
 }
 
 #' Retrieve table metadata
 #'
-#' @param project project containing this table
-#' @param dataset dataset containing this table
+#' @inheritParams insert_dataset
 #' @param table name of the table
 #' @seealso API documentation:
 #'  \url{https://developers.google.com/bigquery/docs/reference/v2/tables/get}
@@ -44,13 +57,14 @@ list_tables <- function(project, dataset, max_results = NULL) {
 #' @export
 #' @examples
 #' \dontrun{
-#' get_table("publicdata", "samples", "natality")
-#' get_table("githubarchive", "github", "timeline")
+#' str(get_table("publicdata", "samples", "natality"))
+#' str(get_table("publicdata", "samples", "gsod"))
+#' str(get_table("githubarchive", "github", "timeline"))
 #' }
 #'
-#' @description \code{get_table} returns a table's metadata as a nested list.
-#'   In addition to a regular error, the condition \code{bigrquery_notFound}
-#'   (which can be handled via \code{\link[base]{tryCatch}})
+#' @description `get_table` returns a table's metadata as a nested list.
+#'   In addition to a regular error, the condition `bigrquery_notFound`
+#'   (which can be handled via [base::tryCatch()])
 #'   is raised if the table could not be found.
 get_table <- function(project, dataset, table) {
   assert_that(is.string(project), is.string(dataset), is.string(table))
@@ -61,8 +75,8 @@ get_table <- function(project, dataset, table) {
 
 #' @rdname get_table
 #' @export
-#' @description \code{exists_table} merely checks if a table exists, and returns
-#'   either \code{TRUE} or \code{FALSE}.
+#' @description `exists_table` merely checks if a table exists, and returns
+#'   either `TRUE` or `FALSE`.
 exists_table <- function(project, dataset, table) {
   tryCatch(
     !is.null(get_table(project = project, dataset = dataset, table = table)),
@@ -82,10 +96,7 @@ exists_table <- function(project, dataset, table) {
 #' get_table("publicdata", "samples", "natality")
 #' }
 delete_table <- function(project, dataset, table) {
-  assert_that(is.string(project), is.string(dataset), is.string(table))
-
-  url <- sprintf("projects/%s/datasets/%s/tables/%s", project, dataset, table)
-  bq_delete(url)
+  bq_delete(bq_path(project, dataset, table))
 }
 
 validate_table_reference <- function(reference) {
@@ -94,37 +105,42 @@ validate_table_reference <- function(reference) {
 }
 
 as_bigquery_table_reference <- function(reference) {
-  list(projectId = reference$project_id,
-       datasetId = reference$dataset_id,
-       tableId = reference$table_id)
+  list(
+    projectId = reference$project_id,
+    datasetId = reference$dataset_id,
+    tableId = reference$table_id
+  )
 }
 
 merge_table_references <- function(partial, complete) {
-  list(project_id = partial$project_id %||% complete$project_id,
-       dataset_id = partial$dataset_id %||% complete$dataset_id,
-       table_id = partial$table_id)
+  list(
+    project_id = partial$project_id %||% complete$project_id,
+    dataset_id = partial$dataset_id %||% complete$dataset_id,
+    table_id = partial$table_id
+  )
 }
 
 #' Copy one or more source tables to a destination table.
 #'
 #' Each source table and the destination table should be table references, that
-#' is, lists with exactly three entries: \code{project_id}, \code{dataset_id},
-#' and \code{table_id}.
+#' is, lists with exactly three entries: `project_id`, `dataset_id`,
+#' and `table_id`.
 #'
 #' @param src either a single table reference, or a list of table references
 #' @param dest destination table
 #' @param project project ID to use for the copy job. defaults to the project of
 #'   the destination table.
 #' @param create_disposition behavior for table creation if the destination
-#'   already exists. defaults to \code{"CREATE_IF_NEEDED"},
-#'   the only other supported value is \code{"CREATE_NEVER"}; see
+#'   already exists. defaults to `"CREATE_IF_NEEDED"`,
+#'   the only other supported value is `"CREATE_NEVER"`; see
 #'   \href{https://cloud.google.com/bigquery/docs/reference/v2/jobs#configuration.copy.createDisposition}{the API documentation}
 #'   for more information
 #' @param write_disposition behavior for writing data if the destination already
-#'   exists. defaults to \code{"WRITE_EMPTY"}, other possible values are
-#'   \code{"WRITE_TRUNCATE"} and \code{"WRITE_APPEND"}; see
+#'   exists. defaults to `"WRITE_EMPTY"`, other possible values are
+#'   `"WRITE_TRUNCATE"` and `"WRITE_APPEND"`; see
 #'   \href{https://cloud.google.com/bigquery/docs/reference/v2/jobs#configuration.copy.writeDisposition}{the API documentation}
 #'   for more information
+#' @inheritParams insert_dataset
 #' @seealso API documentation:
 #'   \url{https://cloud.google.com/bigquery/docs/tables#copyingtable}
 #' @export
@@ -140,7 +156,8 @@ merge_table_references <- function(partial, complete) {
 copy_table <- function(src, dest,
                        create_disposition = "CREATE_IF_NEEDED",
                        write_disposition = "WRITE_EMPTY",
-                       project = NULL) {
+                       project = NULL,
+                       ...) {
   if (validate_table_reference(src)) {
     src <- list(src)
   } else if (!all(vapply(src, validate_table_reference, TRUE)) ||
@@ -151,16 +168,18 @@ copy_table <- function(src, dest,
     stop("dest must be a table reference")
   }
   project <- project %||% dest$project_id
-  url <- sprintf("projects/%s/jobs", project)
+  url <- bq_path(project, jobs = "")
   body <- list(
-      projectId = project,
-      configuration = list(
-          copy = list(
-              sourceTables = lapply(src, as_bigquery_table_reference),
-              destinationTable = as_bigquery_table_reference(dest),
-              createDisposition = create_disposition,
-              writeDisposition = write_disposition
-          )))
+    projectId = project,
+    configuration = list(
+      copy = list(
+        sourceTables = lapply(src, as_bigquery_table_reference),
+        destinationTable = as_bigquery_table_reference(dest),
+        createDisposition = create_disposition,
+        writeDisposition = write_disposition
+      )
+    )
+  )
 
-  bq_post(url, body)
+  bq_post(url, body = bq_body(...))
 }