+++ tests/testthat/test-tables.r
@@ -1,5 +1,33 @@
 context("tables")
 
+
+
+test_that("can create and list tables", {
+  ds <- dataset_10_tables()
+  expect_equal(
+    list_tables(ds$project, ds$dataset),
+    sort(paste0("table", 1:10))
+  )
+})
+
+test_that("can control pagination of list_tables", {
+  tables <- sort(paste0("table", 1:10))
+  ds <- dataset_10_tables()
+
+  expect_equal(list_tables(ds$project, ds$dataset, page_size = 5), tables)
+  expect_equal(list_tables(ds$project, ds$dataset, page_size = 10), tables)
+
+  expect_equal(
+    list_tables(ds$project, ds$dataset, page_size = 1, max_pages = 1),
+    tables[1]
+  )
+  expect_equal(
+    list_tables(ds$project, ds$dataset, page_size = 1, max_pages = 2),
+    tables[1:2]
+  )
+
+})
+
 test_that("table references are validated correctly", {
   valid <- list(project_id = "a", dataset_id = "b", table_id = "c")
   expect_that(validate_table_reference(valid), is_true())